+++ CHANGES.md
@@ -3,6 +3,12 @@
 LogTape changelog
 =================
 
+Version 0.6.0
+-------------
+
+To be released.
+
+
 Version 0.5.0
 -------------